vending machine 퍼블리싱 작업중 1
